Move tests to relevant file
